improvement: expand unique operationId coverage to OpenAPI 3
