+++ src/plugins/validate-semantic/index.js
@@ -10,13 +10,13 @@ import * as schemaValidateActions from "./validators/schema"
 import * as pathsValidateActions from "./validators/paths"
 import * as securityValidateActions from "./validators/security"
 import * as parametersValidateActions from "./validators/parameters"
-import * as operationsValidateActions from "./validators/operations"
 import * as operationsOAS3ValidateActions from "./validators/oas3/operations"
 import * as refsOAS3ValidateActions from "./validators/oas3/refs"
 import * as refs2and3ValidateActions from "./validators/2and3/refs"
 import * as parameters2and3ValidateActions from "./validators/2and3/parameters"
 import * as paths2and3ValidateActions from "./validators/2and3/paths"
 import * as schemas2and3ValidateActions from "./validators/2and3/schemas"
+import * as operations2and3ValidateActions from "./validators/2and3/operations"
 
 export default function SemanticValidatorsPlugin({getSystem}) {
 
@@ -53,7 +53,7 @@ export default function SemanticValidatorsPlugin({getSystem}) {
           ...pathsValidateActions,
           ...securityValidateActions,
           ...parametersValidateActions,
-          ...operationsValidateActions,
+          ...operations2and3ValidateActions,
           ...refs2and3ValidateActions,
           ...operationsOAS3ValidateActions,
           ...refsOAS3ValidateActions,
